mypy fix

+++ stratified_models/tests/quadratic_test.py
@@ -9,7 +9,7 @@ def test_explicit_quadratic() -> None:
     a = np.arange(m**2).reshape((m, m))
     a = a * a.T
     q = MatrixBasedLinearOperator(a)
-    c = np.arange(m)
+    c = np.arange(m, dtype=np.float64)
     d = -170
     f = ExplicitQuadraticFunction(
         q=q,
